feat: add get all plans query

+++ src/main/java/com/pe/platform/payment/infrastructure/persistence/jpa/SubscriptionRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface SubscriptionRepository extends JpaRepository<Subscription, Integer> {
+public interface SubscriptionRepository extends JpaRepository<Subscription, Long> {
     Optional<Subscription> findByProfileId(Long Id);
